Related Records Portals: Change rows count to min and max counts.
* glom/glom_document.dtd: Adapted.
* glom/libglom/data_structure/layout/layoutitem_portal.[h|cc]:
set_rows_count(): Return a min and max, and change it from double to
gulong.
* glom/libglom/document/document.cc: load_after(), save_before():
Store it in the document. The old XML attribute was new in Glom 1.19
anyway, so we can change it.

* ui/developer/window_data_layout.glade: Add an extra SpinButton,
adjust the labels, and add explanatory tooltips.
* glom/mode_design/layout/dialog_layout_details.[h|cc]: Adapted.

* glom/mode_data/db_adddel/db_adddel.[h|cc]: set_height_rows():
Take the min and max.
construct_specified_column(): Increase the size if appropriate when
there are more database rows to show.
* glom/mode_design/layout/dialog_layout_list_related.cc:
* glom/print_layout/print_layout_utils.cc:
* glom/libglom/db_utils.cc:
* glom/mode_data/box_data_list_related.cc: Adapted.

+++ glom/libglom/db_utils.cc
@@ -864,7 +864,7 @@ type_vec_fields get_fields_for_table_from_database(const Glib::ustring& table_na
 
       //Examine each field:
       guint row = 0;
-      const guint rows_count = data_model_fields->get_n_rows();
+      const gulong rows_count = data_model_fields->get_n_rows();
       while(row < rows_count)
       {
         Glib::RefPtr<Gnome::Gda::Column> field_info = Gnome::Gda::Column::create();
